Move MockChainQueryApi/NodeApi out of BaseWalletTest
